requirements.txt
flask==3.0.0
sqlalchemy==2.0.23
apscheduler==3.10.4
requests==2.31.0
resend==2.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
